chore: rename to nbcrypt
