+++ include/nbcrypt.h
@@ -1,9 +1,9 @@
-#ifndef BCRYPT_H
-#define BCRYPT_H
+#ifndef N_BCRYPT_H
+#define N_BCRYPT_H
 
 #include <string>
 
-namespace bcrypt {
+namespace n_bcrypt {
 
     std::string generateHash(const std::string & password , unsigned rounds = 10 );
 
@@ -11,4 +11,4 @@ namespace bcrypt {
 
 }
 
-#endif // BCRYPT_H
+#endif // N_BCRYPT_H
