+++ test/main.cpp
@@ -1,4 +1,4 @@
-#include "bcrypt.h"
+#include "nbcrypt.h"
 #include <iostream>
 #include <string>
